Control.Parse

Description

This module includes a parseMove method, which takes a string and returns a Move (of those defined in the Move module) that matches. If no move matches, Nothing is returned. Functions are also provided to turn two possibly-valid coordinates (as strings) into a Location, and to turn a possibly-valid string into the value of a new Square. These methods need to parse raw user input, so they're rather strict on what they accept as valid input.

Documentation

parseMove :: String -> Maybe Move

Parse a move

parseSquare :: String -> Maybe Square

Turn an Int into a square (or nothing). 0 is turned into Empty.

parseLocation :: String -> String -> Maybe Location

Turn two ints into a Location