Flag

Description

This module describes the flags permitted for the Sudoku application. The flags exist as a data type, and descriptions and long and short option flags are provided as well.

Classes

data Flag

Datatype for command-line flags.

Constructors

Graphical
Enable graphical mode - --graphical, -g

Help
Show help - --help, -h

Version
Show version information - version, -V

File String
Provide a file to load from - --file, -f

Option Descriptions

options :: [OptDescr Flag]

Describe all allowed command-line options.