GTK.GUI

Description

This module contains a datatype for a Sudoku application GUI, which includes everything I think is necessary for a reasonable Sudoku application. It also includes a method for initializing a GUI, a method for binding all of the elements together, and exports the mainWin field so code importing this module can actually use the GUI.

Classes

data GUI

GUI datatype for a Sudoku window.

Constructors

initSudokuGUI :: IO GUI

Initialize a Sudoku GUI and setup GTK+.

Exported Record Entries

mainWin :: GUI -> Window

Main window.