Main

Documentation

repl :: (SudokuBoard -> Move -> Either MoveError SudokuBoard) -> (String -> Maybe Move) -> SudokuBoard -> IO ()

Control loop for Sudoku game. Receive user input, process it (handling errors) and repeat.

maybeDie :: Either MoveError SudokuBoard -> IO ()

Die if we received a QuitError (by using checkError to check the error).

checkError :: MoveError -> IO ()

Die if passed a QuitError, otherwise do nothing.

move :: SudokuBoard -> Move -> Either MoveError SudokuBoard

Set value to board. If an error occurs, the value will not be set. If no error occurs, the requested value will be set.

Try to process a move.

exit :: IO ()

Exit successfully.

die :: IO ()

Exit unsuccessfully.

header :: String

Usage header, printed when improper command-line arguments are given.

versionNum :: String

Version number of game.

version :: String

Full version string, including version numbert.

continue :: IO ()

Do nothing and return an IO ().

commandParse :: ([Flag], [String], [String]) -> IO ([Flag], [String])

Parse command-line arguments

main :: IO ()

Main function, runs application.