Sudoku.Data.Board

Description

This module exports the methods to create a SudokuBoard, the types used to construct one, and all of the methods used to interact with one once it's created.

Classes

data SqVal

Value of a Square, can be 1-9

data Square

Type for Sudoku square value. Stores value and whether this square was placed at creation of the board or not.

data SudokuBoard

Type for Sudoku board. Just a 2D list of squares.

data Location

Location on a SudokuGrid - a coordinate pair.

Methods

isEmpty :: Square -> Bool

Return whether Square is Empty or not.

Location, Square Generation

toLocation :: String -> String -> Maybe Location

Create location from two strings specifying coordinates.

tupleToLocation :: (Int, Int) -> Location

Convert tuple of ints to Location.

toSquare :: Bool -> String -> Maybe Square

Convert Int to Square (or fail and return Nothing), with original flag as specified.

Board Generation

emptyBoard :: SudokuBoard

Create an empty SudokuBoard.

attemptLoad :: String -> SudokuBoard

Attempts to load a board from a file. If it fails, an empty board is returned.

Printing

prettyPrint :: SudokuBoard -> String

Prints a SudokuBoard nicely.

Board Modification

getBoardValue :: SudokuBoard -> Location -> Square

Get the value in a particular square of a Sudoku board.

setBoardValue :: SudokuBoard -> Location -> Square -> SudokuBoard

Returns a SudokuBoard with the value at the two indices modified. Original values are not modified.

eraseBoardValue :: SudokuBoard -> Location -> SudokuBoard

Erase square of a SudokuBoard if it isn't original.

resetBoard :: SudokuBoard -> SudokuBoard

Reset board. Values placed when the board was created are not changed.

checkBoard :: SudokuBoard -> [Location]

Checks if a SudokuBoard is currently valid or not. Checks all rows, columns, and subgrids with checkList and returns a list of squares in error.