Sudoku.Data.Board.Internal

Description

This module includes the methods to create a SudokuBoard, the types used to construct one, and all of the methods used to interact with one once it's created.

Documentation

data Location

Location on a SudokuGrid - a coordinate pair.

Constructors

Loc Int Int
Create location given row and column.

data Square

Type for Sudoku square value. Stores value and whether this square was placed at creation of the board or not.

Constructors

Empty
An empty square.

Val

Fields

value :: SqVal

Value of square.

isOrig :: Bool

Did this square come with the board?

data SqVal

Value of a Square, can be 1-9

Constructors

V1

V2

V3

V4

V5

V6

V7

V8

V9
Constructors for values 1-9.

data SudokuBoard

Type for Sudoku board. Just a 2D list of squares.

Constructors

SudokuBoard [[Square]]
Create board from 2D list.

isEmpty :: Square -> Bool

Return whether Square is Empty or not.

toSquare :: Bool -> String -> Maybe Square

Convert Int to Square (or fail and return Nothing), with original flag as specified.

squareGen :: Bool -> SqVal -> Maybe Square

Helper method for toSquare to cut down on code repetition.

toLocation :: String -> String -> Maybe Location

Create location from two strings specifying coordinates.

tupleToLocation :: (Int, Int) -> Location

Convert tuple of ints to Location.

emptyBoard :: SudokuBoard

Create an empty SudokuBoard.

attemptLoad :: String -> SudokuBoard

Attempts to load a board from a file. If it fails, an empty board is returned.

resetBoard :: SudokuBoard -> SudokuBoard

Reset board. Values placed when the board was created are not changed.

eraseLocation :: Square -> Square

Erase square if it isn't an original value.

getBoardValue :: SudokuBoard -> Location -> Square

Get the value in a particular square of a Sudoku board.

setBoardValue :: SudokuBoard -> Location -> Square -> SudokuBoard

Returns a SudokuBoard with the value at the two indices modified. Original values are not modified.

eraseBoardValue :: SudokuBoard -> Location -> SudokuBoard

Erase square of a SudokuBoard if it isn't original.

checkBoard :: SudokuBoard -> [Location]

Checks if a SudokuBoard is currently valid or not. Checks all rows, columns, and subgrids with checkList and returns a list of squares in error.

prettyPrint :: SudokuBoard -> String

Prints a SudokuBoard nicely.

toIndex :: String -> Maybe Int

Turn an int into an index value.

createBoard :: [[String]] -> SudokuBoard

Create board from array of array of string. Used for loading from files.

checkLocGrid :: [[Location]] -> SudokuBoard -> [Location]

Check a list of lists of squares for duplicates

checkRows :: SudokuBoard -> [Location]

Returns a list of all locations that are invalid (for rows).

checkCols :: SudokuBoard -> [Location]

Returns a list of all locations that are invalid (for columns).

checkSubgrids :: SudokuBoard -> [Location]

Returns a list of all locations that are invalid (for subgrids).

checkList :: [(Location, Square)] -> [Location]

Takes a list of tuples of Int and Square. The Int is an indexing tag. Returns a list of locations corresponding to matching squares.

checkPair :: Square -> Square -> Bool

Check if a pair of Squares are equal.

zipLists :: [[a]] -> [[b]] -> [[(a, b)]]

Helper function for check functions. Given two lists of lists, returns a list of lists of pairs.

subgridHelper :: Int -> [Location]

Given an Int indicating the index of a subgrid, returns a list of Locations matching that subgrid. TODO find a better way to do this.

sudokuSpacer :: a -> [a] -> [a]

Given an array and a spacer, uses the spacer to separate the first three from the middle three rows and the middle three from the last three rows.