Util.Other

Description

This module houses any methods that don't quite fit in anywhere else (or are generic enough to deserve placement in a separate module). Currently, there are two methods for incrementing tuples, two methods for generating lists of IO objects, and a method for converting a string into an integer conditionally.

Tuples

addToTupleIf :: Int -> (Int -> Bool) -> (Int -> Bool) -> (Int, Int) -> (Int, Int)

Add int to tuple of int if the first condition is true for the first element of the tuple and the second is true for the second element.

incrementTupleIf :: (Int -> Bool) -> (Int -> Bool) -> (Int, Int) -> (Int, Int)

Special case for adding one to tuple.

IO

getIORow :: Int -> IO a -> IO [a]

Given an Int n and a function returning an IO a, return an IO list of n a.

getIOGrid :: Int -> IO a -> IO [[a]]

Given an Int n and a function returning an IO a, return an IO list of lists of n a.

String

sToIntRange :: String -> [Int] -> Maybe Int

Given a string and a range, return either Just the string converted to an Int if it is in range, or Nothing.

List

randomElem :: Show a => [a] -> a

Given a list, return a random element from that list.

getElem :: [[a]] -> (Int, Int) -> a

Get the value at the provided indices from a 2D list.

setElem :: [[a]] -> (Int, Int) -> a -> [[a]]

Set the value at the provided indicies in a 2D list.

setElemTuple :: Show a => [[a]] -> ((Int, Int), a) -> [[a]]

Identitical to setElem, but take the value and indices as a tuple.