View

Description

This module exists because I'm trying to follow the MVC paradigm with this application. This module wraps all of the GUI stuff together, so the controller can simply call these two methods to start the GUI if the user asks for it. Currently, there seems to be no way to control the GUI from the controller once it has started, so this approach might need to be changed.

Documentation

initSudokuView :: IO Window

Initialize Sudoku view

runSudokuWindow :: WindowClass window => window -> IO ()

Run GUI window.