Main

Description

This module provides several things. First, it provides a control loop for receiving and interpreting player input. This loop takes input, attempts to parse it (using Parse), gets the Move (using Move) the player intended, applies it, and either loops with the new state of the board after the move (if the move was valid), or loops with the old state of the board (if there was an error). It also shows the board on each iteration and kills the program if a QuitError comes up (if the player asked to quit).

There are also functions for each possible move, to actually apply the player's moves. If the moves fail for any reason, these functions return errors instead of valid new boards.

The rest of the module is code to handle command line arguments, some strings holding info about the program, and the actual Main function. The main function parses arguments, prints some stuff out, and kickstarts the control loop.

Control

repl :: (SudokuBoard -> Move -> Either MoveError SudokuBoard) -> (String -> Maybe Move) -> SudokuBoard -> IO ()

Control loop for Sudoku game. Receive user input, process it (handling errors) and repeat.

move :: SudokuBoard -> Move -> Either MoveError SudokuBoard

Try to process a move.

Command-Line Processing

commandParse :: ([Flag], [String], [String]) -> IO ([Flag], [String])

Parse command-line arguments

exit :: IO ()

Exit and indicate success.

die :: IO ()

Exit and indicate failure.

continue :: IO ()

Do nothing and return an IO ().

Program Information.

header :: String

Usage header, printed when improper command-line arguments are given.

version :: String

Full version string, including version numbert.

Main

main :: IO ()

Main function, runs application.