Util.Other

Description

This module houses any methods that don't quite fit in anywhere else (or are generic enough to deserve placement in a separate module). Currently, there are two methods for incrementing tuples, two methods for generating lists of IO objects, and a method for converting a string into an integer conditionally.

Tuples

addToTupleIf :: Int -> (Int -> Bool) -> (Int -> Bool) -> (Int, Int) -> (Int, Int)

Add int to tuple of int if the first condition is true for the first element of the tuple and the second is true for the second element.

incrementTupleIf :: (Int -> Bool) -> (Int -> Bool) -> (Int, Int) -> (Int, Int)

Special case for adding one to tuple.

Monads

getMonadicGrid :: Monad m => Int -> m a -> m [[a]]

Given an Int n and a function returning a monad a, return a monad list of lists of n a.

Strings

sToIntRange :: String -> [Int] -> Maybe Int

Given a string and a range, return either Just the string converted to an Int if it is in range, or Nothing.